dubai_rent_bot/backend/requirements.txt
fastapi
uvicorn[standard]
aiogram==3.0.0b7
asyncpg
databases[postgresql]
pydantic
python-dotenv
httpx